# swagger.yaml
openapi: 3.0.0
info:
  title: Random Image API
  contact: {}
  version: '1.0'
servers:
  - url: https://randoImg.onrender.com/api/v1
    variables: {}
paths:
...
  /images:
    post:
      security:
        - cookieAuth: []
      tags:
        - Images
      summary: Create Image
      operationId: CreateImage
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageRequest'
            example:
              name: 'My Image'
              imgageURL: '/mike.jpeg'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      security: []
      tags:
        - Images
      summary: Get All Images
      operationId: GetAllImages
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    CreateImageRequest:
      title: CreateImageRequest
      required:
        - name
        - imageURL
      type: object
      properties:
        name:
          type: string
        imageURL:
          type: string
      example:
        name: My Image
        imageURL: /mike.jpeg
  securitySchemes:
    cookieAuth:         # arbitrary name for the security scheme; will be used in the "security" key later
      type: apiKey
      in: cookie
      name: token  # cookie name 
security:
  - cookieAuth: []
...